Added temporary Automatic DB handler for Ecoli kmerAligner and polished kmeraligner rule (#12)
* Added temporary Automatic DB handler for Ecoli kmerAligner and polished kmeraligner rule

Current Ecoli kmeraligner database is hosted as part of a repo, to simplify, only the db file is downloaded using wget (thus wget has been added to envs/kmeraligner.yaml). The repo is prob not stable (not maintained) and thus this fix should be seen as a temporary addition to #7. To make rule comply with latest Logging and Messaging setup, the kmeraligner rule have recieved a polishing.

* Amend: Replaced wget with Curl

Since firewalls seems to block wget, Curl is used as a replacement to wget.