V0.5 (#39)
* Fixed the config files and the Snakemake + rules to account for those changes

* The conda options for plasmidfinder and resfinder are back to general
and not point to a specific env

* Fixed paths

* Fixed all the envs yaml file

* Deactivated kmerfinder rule

* Relative config paths and configfile added in Snakefile

Part of requirements for resolving #5

* Added examples

Added a public E. coli sample as well as a Actinobacillus pleuropneumoniae sample to test the E. coli pipe as well as the pipe for unspecified samples. Addresses #6

* Getting rid of Large File Storage dep

* Fixed some path on configfile
Added the resources folder to gitignore
Minor fixes in workflow/

* Added AMRFinderPlus (#8)

Added AMRfinderplus, requires preloaded DB pointed out by the Species_config file

* Removed message, polished comments and added organisms support

Addressed issues with pull request #8. Organism support is a bit yanky now though, as you would be expected to provide the --organism XYZ directly into the species specific config file. Alternately you would have to leave a blank string ('')

* Automatic dbs (#9)

* Added rules for automatically building CGEFinders databases. Added logging to CGEfinder rules

Databases for CGEfinders are now automatically downloaded and indexed with KMA's if their db folder is missing. Logging has been set up to enhance debugging and usage information. This commit contributes to #5 #6 and #7

* Prepared for automatic AMRFinderDB

Premerge automatic DB of AMRFinder prior to implementation

* Added AMRfinder DB and polished messaging

With the addition of AMRfinderPlus, Automatic Database handling has to come along. Now the database is automatically installed if missing. In addition messages are now named after their rule names, to enhance debugging

* Corrected log positions, Amended AMRFinder db setup to names

Logs are now placed into subcategory folders (based on sample or according to subject e.g. Databases

* Cleaning: Removed redundant tools KMERFinder & Resistance Genes Detection (#10)

Some tools are not used in practise any more and are thus redundant. Objects related to KMERFinder & Resistance Genes detection rules have been purged.

Co-authored-by: SimoneScrima <simonescrima@gmail.com>

* Added MLST By Torsten Seeman (#11)

MLST has been added and formatted to fit with the Logging conventions and Messaging convensions of #9

---------

Co-authored-by: SimoneScrima <simonescrima@gmail.com>

* Added temporary Automatic DB handler for Ecoli kmerAligner and polished kmeraligner rule (#12)

Current Ecoli kmeraligner database is hosted as part of a repo, to simplify, only the db file is downloaded using wget (thus wget has been added to envs/kmeraligner.yaml). The repo is prob not stable (not maintained) and thus this fix should be seen as a temporary addition to #7. To make rule comply with latest Logging and Messaging setup, the kmeraligner rule have recieved a polishing.

* Amend: Replaced wget with Curl

Since firewalls seems to block wget, Curl is used as a replacement to wget.

* Force update AMRFinder setup

AMRfinder_update results in an error if database for some reason is not completely setup. Thus, everytime the setup rule is executed, it MUST overwrite the preexisting (if it's the same version) to mitigate future setup errors.
Also corrected EcoliKmerAligner message.

* Removed tasks not involved in routine

While these tasks are NOT discarded on the long run, excluding these makes code more readable, while retaining the enhancements in commit history.

* Updated README - Pull FIRST

Make sure to update the supported tools section in all future pull requests, when relevant

* Removed unusued scripts and envs as well

* Added SerotypeFinder database creation (#14)

SerotypeFinder database is now generated automatically (if missing). Added messaging and logging for serotypefinder rule.

* Added update rule for MLST (#15)

Due to some limitations in MLST setup, database updates can't be tied to snakemake rules output, and thus made as a dependency of the MLST execution. To manually force update (TAKES TIME\!) run

* Update db_setups.smk

Added time warning on the update_MLST rule

* Added database creation file

The database creation file is now chained into the MLST run, ensuring that if the database creation file does not exist, the DB is build

* Removed multithreading since resulted in error in dowloading the database

---------

Co-authored-by: SimoneScrima <simonescrima@gmail.com>

* Added create.date files for each db_setup rules (#19)

The file is a simple way of investegating the creation date for each used databases in the pipe.

* Updated readme documentation (#34)

* Changed the order of sections in the README.md

* Added description of which files to change when including species

* Update README.md

* Removed abstract and TCO from TOC

---------

Co-authored-by: Kasper Thystrup Karstensen <kathka@gmail.com>

* Strict minimization - Deletions (#29)

Removed a lot of files relating to NBDev, to enhance the overview of the repository. The removed files are not used in the later development of this pipe.

* Removed additional samplesheet

* Fixed config and snakemake following issue #26 suggestion (#37)

Validated on Ubuntu

* Removed unsupported rules

For stability, unsupported rules where removed prior to pre-release. Rules corresponds to #38, #40, and #41

* Polished ResFinder command

Resfinder command is no longer a python module call, rather an executionable command, altså implicitly stated location of the KMerAlinger database

---------

Co-authored-by: SimoneScrima <simonescrima@gmail.com>
Co-authored-by: Rasmus Amund Henriksen <36986552+RAHenriksen@users.noreply.github.com>
Co-authored-by: Thor Bech Johannesen <thej@ssi.dk>